fix: otp-1715: MetadataImportService.findById uses input as long instead of string

+++ grails-app/controllers/de/dkfz/tbi/otp/ngsdata/MetadataImportController.groovy
@@ -176,7 +176,7 @@ class MetadataImportController implements CheckAndCall, PlainResponseExceptionHa
 
     def multiDetails() {
         List<MetaDataFileWrapper> metaDataFiles = params.metaDataFiles.collect {
-            MetaDataFile file = metadataImportService.findById(it)
+            MetaDataFile file = metadataImportService.findById(it as long)
 
             new MetaDataFileWrapper(
                     metaDataFile: file,
